# home-companion-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Companion phone API",
    "version": "0.1.0",
    "description": "Shared household API. Phone tokens use hs_sess_; module endpoints also accept scoped hs_key_ keys. Cookies never authenticate this API. Keys and admin operations are web-only. Complete bootstrap password change on the web before phone login."
  },
  "servers": [
    {
      "url": "https://home-companion.home.arpa/api/v1"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Loopback development only"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/groceries/stock": {
      "get": {
        "summary": "Search product stock",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          }
        ]
      },
      "post": {
        "summary": "Add existing/new product packages atomically",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStock"
              }
            }
          }
        }
      }
    },
    "/groceries/stock/{id}/adjust": {
      "post": {
        "summary": "Apply signed package adjustment; never below zero",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Adjustment"
              }
            }
          }
        }
      }
    },
    "/groceries/items": {
      "get": {
        "summary": "List item identities",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      },
      "post": {
        "summary": "Create/reuse an item, including items with no products",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemInput"
              }
            }
          }
        }
      }
    },
    "/groceries/desired": {
      "get": {
        "summary": "List all targets with current totals",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shopping"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
...
    "/groceries/desired/{id}": {
      "put": {
        "summary": "Set or replace an item target",
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Target"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove an item target",
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque hs_sess_ or scoped hs_key_"
      }
    },
    "schemas": {
...
      "ItemInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "kind"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "kind": {
            "type": "string",
            "enum": ["COUNT", "MASS", "VOLUME"]
          },
          "category": {
            "type": "string"
          }
        }
      },
      "Item": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": ["COUNT", "MASS", "VOLUME"]
          },
          "category": {
            "type": "string"
          }
        }
      },
      "NewProduct": {
        "type": "object",
        "additionalProperties": false,
        "required": ["item", "kind", "packageQuantity", "unit"],
        "properties": {
          "item": {
            "type": "string",
            "maxLength": 80
          },
          "kind": {
            "type": "string",
            "enum": ["COUNT", "MASS", "VOLUME"]
          },
          "category": {
            "type": "string"
          },
          "brand": {
            "type": ["string", "null"],
            "maxLength": 80
          },
          "attributes": {
            "type": "array",
            "maxItems": 10,
            "items": {
              "type": "string",
              "maxLength": 80
            }
          },
          "packageQuantity": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999",
            "exclusiveMinimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          },
          "unit": {
            "type": "string",
            "enum": ["unit", "g", "kg", "ml", "l"]
          }
        }
      },
      "AddStock": {
        "type": "object",
        "additionalProperties": false,
        "required": ["packages"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "newProduct": {
            "$ref": "#/components/schemas/NewProduct"
          },
          "packages": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999",
            "exclusiveMinimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          }
        },
        "oneOf": [
          {
            "required": ["productId"],
            "not": {
              "required": ["newProduct"]
            }
          },
          {
            "required": ["newProduct"],
            "not": {
              "required": ["productId"]
            }
          }
        ]
      },
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "item": {
            "type": "string"
          },
          "brand": {
            "type": ["string", "null"]
          },
          "attributes": {
            "type": "string"
          },
          "packageQuantity": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999",
            "exclusiveMinimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          },
          "unit": {
            "type": "string",
            "enum": ["unit", "g", "kg", "ml", "l"]
          },
          "packages": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "lastPurchasedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "productKey": {
            "type": "string"
          }
        }
      },
      "Stock": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "ShoppingRow": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "item": {
            "type": "string"
          },
          "target": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999",
            "exclusiveMinimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          },
          "current": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999"
          },
          "needed": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999"
          },
          "unit": {
            "type": "string",
            "enum": ["unit", "g", "kg", "ml", "l"]
          },
          "suggestedProduct": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "packagesToBuy": {
            "type": ["number", "null"]
          },
          "noProductKnown": {
            "type": "boolean"
          }
        }
      },
      "Shopping": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingRow"
            }
          }
        }
      },
      "Target": {
        "type": "object",
        "additionalProperties": false,
        "required": ["quantity", "unit"],
        "properties": {
          "quantity": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999",
            "exclusiveMinimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          },
          "unit": {
            "type": "string",
            "enum": ["unit", "g", "kg", "ml", "l"]
          },
          "preferredBrand": {
            "type": "string"
          }
        }
      },
      "Adjustment": {
        "type": "object",
        "additionalProperties": false,
        "required": ["delta"],
        "properties": {
          "delta": {
            "type": "number",
            "description": "Exact decimal, at most 3 stored fractional digits; absolute value <= 999999999.999"
          }
        }
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          }
        },
        "description": "RFC 9457 problem; validation=422, identity conflict=409, authentication=401, authorization=403, throttle=429"
      }
    }
  }
}